requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
nltk>=3.8.1
plotly>=5.13.1
wordcloud>=1.8.2.2
google-ads>=21.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
